elimu_hub/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication
PyJWT==2.8.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Vector database
chromadb==0.4.15

# Text processing and embeddings
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2

# PDF processing
PyMuPDF==1.23.8
pytesseract==0.3.10
Pillow==10.1.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and utilities
loguru==0.7.2
typing-extensions==4.8.0

# HTTP client for Ollama
httpx==0.25.2
requests==2.31.0

# Text chunking and processing
langchain-text-splitters==0.0.1
nltk==3.8.1

# Data validation and serialization
email-validator==2.1.0
